git-svn-id: https://svn.jnitinc.com/svn/vsdev/branches/VitalSignsMongo@1062 6b605366-3ce6-ac4f-954a-f758cfb76b34

+++ VitalSignsWeb/VitalSigns.API/Models/Configurator/AdvancedSettingsModel.cs
@@ -11,10 +11,10 @@ namespace VitalSigns.API.Models
 
 
         [JsonProperty("memory_threshold")]
-        public int? MemoryThreshold { get; set; }
+        public double? MemoryThreshold { get; set; }
 
         [JsonProperty("cpu_threshold")]
-        public int? CpuThreshold { get; set; }
+        public double? CpuThreshold { get; set; }
 
         [JsonProperty("server_days_alert")]
         public int? ServerDaysAlert { get; set; }
